f64 cuda tests passing

+++ src/tensor_ops/negate/negate.cu
@@ -6,6 +6,6 @@ UNARY_OP(float, negate_forward_f32, negate_backward_f32, NegateKernelOp,
         -x,
         -1.0)
 
-UNARY_OP(float, negate_forward_f64, negate_backward_f64, NegateKernelOp,
+UNARY_OP(double, negate_forward_f64, negate_backward_f64, NegateKernelOp,
         -x,
         -1.0)
